Move upgrade bookmarklet view to assetmgr.views
